convert MotorControllers to a class
